[common] Move logging::get_disk_sink to new header
Deprecate the use of its prior location and its use when spdlog is
disabled.
