+++ common/BUILD.bazel
@@ -66,6 +66,7 @@ drake_cc_package_library(
         ":sorted_pair",
         ":string_container",
         ":temp_directory",
+        ":text_logging_spdlog",
         ":timer",
         ":type_safe_index",
         ":unused",
@@ -111,12 +112,26 @@ drake_cc_library(
         "never_destroyed.h",
         "ssize.h",
         "text_logging.h",
+        # Remove with deprecation on 2026-03-01.
+        "text_logging_spdlog.h",
     ],
     deps = [
         ":fmt",
         "@eigen",
         "@fmt",
         "@spdlog",
+    ],
+)
+
+drake_cc_library(
+    name = "text_logging_spdlog",
+    hdrs = ["text_logging_spdlog.h"],
+    tags = [
+        # Don't lint us twice. (Remove with deprecation on 2026-03-01.)
+        "nolint",
+    ],
+    deps = [
+        ":essential",
     ],
 )
 
@@ -1221,6 +1236,7 @@ drake_cc_googletest(
     use_default_main = False,
     deps = [
         ":essential",
+        ":text_logging_spdlog",
     ],
 )
 
@@ -1235,6 +1251,9 @@ drake_cc_googletest(
         "test/text_logging_test.cc",
         "text_logging.cc",
         "text_logging.h",
+        # Remove both of these with deprecation on 2026-03-01.
+        "drake_deprecated.h",
+        "text_logging_spdlog.h",
     ],
     defines = [
         "TEXT_LOGGING_TEST_SPDLOG=0",
@@ -1255,6 +1274,9 @@ drake_cc_googletest(
         "test/text_logging_ostream_test.cc",
         "text_logging.cc",
         "text_logging.h",
+        # Remove both of these with deprecation on 2026-03-01.
+        "drake_deprecated.h",
+        "text_logging_spdlog.h",
     ],
     defines = [
         "TEXT_LOGGING_TEST_SPDLOG=0",
